+++ modules/core/src/main/java/com/mammb/code/editor/core/model/CsvEditorModel.java
@@ -210,7 +210,7 @@ public class CsvEditorModel implements EditorModel {
     }
 
     @Override
-    public void wrap(double width) {
+    public void wrap(int width) {
         // do nothing
     }
