+++ modules/core/src/main/java/com/mammb/code/editor/core/model/TextEditorModel.java
@@ -433,7 +433,7 @@ public class TextEditorModel implements EditorModel {
     }
 
     @Override
-    public void wrap(double width) {
+    public void wrap(int width) {
         view.wrapWith(width);
     }
